thermal: Use highest temperature when using automatic sensor location
When using an automatic sensor location, the highest temperature of all
sensors found is used. Similarly, the lowest critical temperature is
used when automatic warning levels are enabled. I've been wanting this
feature for a while, since at the moment my laptop's two thermal zones
of the dual core cpu are different by 13°C.

The maximum number of sensors is hardcoded to 10, because I am too lazy
to do memory management. If there are more sensors they are ignored with
an error message.